Fix menu import and restructure code needed to add in TenantUserMenu component in the tenant dashboard.
