--- rentme/src/components/SideMenu/index.js
@@ -1,3 +1,0 @@
-import OwnerUserMenu from "./OwnerUserMenu";
-
-export default OwnerUserMenu;
