+++ rentme/src/components/propertyOwner/AddPropertyForm/index.js
@@ -6,7 +6,7 @@ import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Button from "@material-ui/core/Button";
 import { withStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
-import OwnerUserMenu from "../../SideMenu";
+import OwnerUserMenu from "../../SideMenu/OwnerUserMenu";
 import Icon from "@material-ui/core/Icon";
 import Grid from "@material-ui/core/Grid";
