+++ rentme/src/components/propertyOwner/PropertyCard/index.js
@@ -4,7 +4,7 @@ import Button from "@material-ui/core/Button";
 import Tooltip from "@material-ui/core/Tooltip";
 import axios from "axios";
 import { withStyles } from "@material-ui/core/styles";
-import OwnerUserMenu from "../../SideMenu";
+import OwnerUserMenu from "../../SideMenu/OwnerUserMenu";
 import Icon from "@material-ui/core/Icon";
 
 const drawerWidth = 240;
